+++ src/models/category.cr
@@ -2,7 +2,7 @@ class Category < BaseModel
   skip_default_columns
 
   table do
-    primary_key id : Int32
+    primary_key id : Int16
     column name : String
     column hashtag : String
     column slug : String
